Dev to Main PR to fix all route errors

+++ backend/db/migrations/02-create-spot.js
@@ -25,7 +25,7 @@ module.exports = {
         allowNull: true
       },
       price: {
-        type: Sequelize.DECIMAL(5, 2),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       address: {
@@ -45,11 +45,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.DECIMAL(9, 7),
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       lng: {
-        type: Sequelize.DECIMAL(10, 7),
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       ownerId: {
